# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: LucidFence API
  version: 2.0.0-dev
  description: Geofencing multi-UEM local-first. Toda ruta declara x-capability (public o la capacidad RBAC exigida).
servers:
  - url: http://127.0.0.1:8765
paths:
...
  /api/v1/engine/run-once:
    post:
      x-capability: engine:run
      summary: Ejecuta un ciclo ahora
      responses:
        "200":
          description: Ciclo ejecutado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleStats"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/events:
    get:
      x-capability: device:read
      summary: Transiciones paginadas, de la más reciente a la más antigua
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPage"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/actions:
    get:
      x-capability: device:read
      summary: Acciones paginadas, de la más reciente a la más antigua
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionPage"
        "400":
          $ref: "#/components/responses/Error"
...
  /api/v1/settings/enforcement:
    put:
      x-capability: engine:config
      summary: Cambia solo el bloque de enforcement y lo pone en vigor en el acto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnforcementSettings"
      responses:
        "200":
          description: Ajustes completos ya vigentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/settings/webhooks:
    put:
      x-capability: engine:config
      summary: Cambia solo el bloque de webhook; secret es de solo escritura
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookSettingsUpdate"
      responses:
        "200":
          description: Ajustes completos ya vigentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/settings/egress:
    put:
      x-capability: engine:config
      summary: Cambia solo la allowlist de egress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EgressSettings"
      responses:
        "200":
          description: Ajustes completos ya vigentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/settings/risk:
    put:
      x-capability: engine:config
      summary: Cambia solo el contexto de riesgo (turnos, riesgo por zona y franja nocturna)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskSettings"
      responses:
        "200":
          description: Ajustes completos ya vigentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/settings/validate:
    post:
      x-capability: engine:config
      summary: Comprueba los ajustes y las URLs de salida sin guardar ni enviar nada
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsValidationRequest"
      responses:
        "200":
          description: Veredicto por canal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsValidation"
        "400":
          $ref: "#/components/responses/Error"
components:
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
    Cursor:
      name: cursor
      in: query
      description: Cursor opaco devuelto como next_cursor por la página anterior. Ausente empieza por lo más reciente.
      schema:
        type: string
  responses:
    Error:
      description: Error con forma única
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        detail: {}
      required: [error, code]
...
    ProviderHealth:
      type: object
      properties:
        ok:
          type: boolean
        error:
          type: string
        devices:
          type: integer
        latency_ms:
          type: integer
        last_ok:
          type: string
          format: date-time
      required: [ok, devices, latency_ms]
    EnforcementSettings:
      type: object
      description: >-
        Guardarraíles vigentes del motor (spec §5.4). live_actions nula
        significa "todas las acciones salen en vivo"; una lista vacía deja
        todo en dry-run.
      properties:
        mode:
          type: string
          enum: [observe, enforce]
        live_actions:
          type: [array, "null"]
          items:
            type: string
        allow_wipe:
          type: boolean
        wipe_allowlist:
          type: [array, "null"]
          items:
            type: string
        action_cooldown_seconds:
          type: integer
      required: [mode, live_actions, allow_wipe, wipe_allowlist, action_cooldown_seconds]
    CycleStats:
      type: object
      properties:
        at:
          type: string
          format: date-time
        duration_ms:
          type: integer
        mode:
          type: string
        devices_total:
          type: integer
        inside:
          type: integer
        outside:
          type: integer
        unknown:
          type: integer
        risk_evaluated:
          type: integer
        risk_failed:
          type: integer
        by_severity:
          type: object
          additionalProperties:
            type: integer
        transitions:
          type: integer
        actions_planned:
          type: integer
        actions_executed:
          type: integer
        actions_suppressed:
          type: integer
        actions_blocked:
          type: integer
        incidents_opened:
          type: integer
        incidents_closed:
          type: integer
        alerts_fired:
          type: integer
        handoffs_pending:
          type: integer
        deliveries:
          type: integer
        deliveries_failed:
          type: integer
        evaluation_errors:
          type: integer
        persistence_errors:
          type: integer
        providers:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ProviderHealth"
      required: [at, duration_ms, mode, devices_total, inside, outside, unknown, transitions, actions_planned, actions_executed, actions_suppressed, actions_blocked, incidents_opened, incidents_closed, alerts_fired, handoffs_pending, deliveries, deliveries_failed, evaluation_errors, persistence_errors, providers, risk_evaluated, risk_failed, by_severity]
...
    Transition:
      type: object
      properties:
        at:
          type: string
          format: date-time
        device_id:
          type: string
        device_name:
          type: string
        from:
          type: string
        to:
          type: string
      required: [at, device_id, device_name, from, to]
    ActionResult:
      type: object
      properties:
        adapter:
          type: string
        ok:
          type: boolean
        device_id:
          type: string
        device_name:
          type: string
        action:
          type: string
        params:
          type: object
          additionalProperties: true
        dry_run:
          type: boolean
        simulated:
          type: boolean
        error:
          type: string
        command_id:
          type: string
        note:
          type: string
        at:
          type: string
          format: date-time
        fence_id:
          type: string
        trigger:
          type: string
        route_id:
          type: string
        policy_id:
          type: string
        playbook_id:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        blocked:
          type: boolean
          description: Solo se emite si true; su ausencia no acredita autorización ni ejecución.
        error_type:
          type: string
      required: [adapter, ok, device_id, device_name, action, dry_run, simulated, at]
...
    WebhookSettings:
      type: object
      description: >-
        Canal de webhook firmado. secret_set dice si hay credencial guardada;
        el valor no sale nunca por la API.
      properties:
        url:
          type: string
        format:
          type: string
          enum: [native, ocsf]
        events:
          type: array
          items:
            type: string
            enum: [incident.opened, incident.closed, handoff.pending, action.executed, alert.fired]
        enabled:
          type: boolean
        secret_set:
          type: boolean
      required: [url, format, events, enabled, secret_set]
    WebhookSettingsUpdate:
      type: object
      description: >-
        Cuerpo de PUT /settings/webhooks. secret es de solo escritura: si
        falta no se toca el secreto guardado, la cadena vacía lo borra y
        cualquier otro valor lo sustituye. secret_set no se acepta.
      properties:
        url:
          type: string
        format:
          type: string
          enum: [native, ocsf]
        events:
          type: array
          items:
            type: string
            enum: [incident.opened, incident.closed, handoff.pending, action.executed, alert.fired]
        enabled:
          type: boolean
        secret:
          type: string
          writeOnly: true
      required: [url, format, events, enabled]
    NtfySettings:
      type: object
      properties:
        url:
          type: string
        enabled:
          type: boolean
        token_set:
          type: boolean
      required: [url, enabled, token_set]
    EgressSettings:
      type: object
      description: Allowlist de salida. Vacía deniega todo; allow_private abre RFC1918 y loopback, nunca link-local.
      properties:
        hosts:
          type: array
          items:
            type: string
        allow_private:
          type: boolean
      required: [hosts, allow_private]
    RiskSettings:
      type: object
      properties:
        shift_zones:
          type: object
          additionalProperties:
            type: string
        zone_risk:
          type: object
          additionalProperties:
            type: number
        off_hours_start:
          type: integer
        off_hours_end:
          type: integer
      required: [shift_zones, zone_risk, off_hours_start, off_hours_end]
    Settings:
      type: object
      properties:
        schema_version:
          type: integer
        enforcement:
          $ref: "#/components/schemas/EnforcementSettings"
        webhook:
          $ref: "#/components/schemas/WebhookSettings"
        ntfy:
          $ref: "#/components/schemas/NtfySettings"
        egress:
          $ref: "#/components/schemas/EgressSettings"
        risk:
          $ref: "#/components/schemas/RiskSettings"
        updated_at:
          type: string
          format: date-time
      required: [schema_version, enforcement, webhook, ntfy, egress, risk, updated_at]
    SettingsValidationRequest:
      type: object
      description: >-
        Bloques que el formulario tiene en pantalla. Los ausentes se toman de
        los ajustes guardados. No acepta el secreto del webhook.
      properties:
        enforcement:
          $ref: "#/components/schemas/EnforcementSettings"
        webhook:
          $ref: "#/components/schemas/WebhookSettings"
        ntfy:
          $ref: "#/components/schemas/NtfySettings"
        egress:
          $ref: "#/components/schemas/EgressSettings"
        risk:
          $ref: "#/components/schemas/RiskSettings"
    SettingsValidationChannel:
      type: object
      properties:
        channel:
          type: string
          enum: [webhook, ntfy]
        enabled:
          type: boolean
        url:
          type: string
        ok:
          type: boolean
        reason:
          type: string
        addresses:
          type: array
          items:
            type: string
      required: [channel, enabled, ok]
    SettingsValidation:
      type: object
      properties:
        ok:
          type: boolean
        error:
          type: string
        field:
          type: string
        channels:
          type: array
          items:
            $ref: "#/components/schemas/SettingsValidationChannel"
      required: [ok, channels]
    EventPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transition"
        next_cursor:
          type: string
      required: [items, next_cursor]
    ActionPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ActionResult"
        next_cursor:
          type: string
      required: [items, next_cursor]